+++ utils/command_test.go
@@ -4,6 +4,8 @@ import (
 	"flag"
 	"testing"
 
+	"fmt"
+
 	"github.com/urfave/cli"
 )
 
@@ -81,5 +83,37 @@ func TestRequiresArgsInRangeErrMax(t *testing.T) {
 	err := RequiresArgsInRange(1, 2, ctx)
 	if err == nil {
 		t.Error("expected error: arg count over max")
+	}
+}
+
+func TestCmdHandlerCliErr(t *testing.T) {
+	err := cli.NewExitError("test cli error", 1)
+	e := CmdHandler(err)
+	if e == nil {
+		t.Error("expected error but got nil")
+	}
+	_, ok := e.(cli.ExitCoder)
+	if !ok {
+		t.Error("expected returned error to be of type cli.ExitCoder but was not")
+	}
+}
+
+func TestCmdHandlerRegErr(t *testing.T) {
+	err := fmt.Errorf("test regular error")
+	e := CmdHandler(err)
+	if e == nil {
+		t.Error("expected error but got nil")
+	}
+	_, ok := e.(cli.ExitCoder)
+	if !ok {
+		t.Error("expected returned error to be of type cli.ExitCoder but was not")
+	}
+}
+
+func TestCmdHandlerNoErr(t *testing.T) {
+	var err error
+	e := CmdHandler(err)
+	if e != nil {
+		t.Errorf("expected nil, but got %v", e)
 	}
 }
